[fork] Just fix all the compiler warnings
- Make config directory creation cross-platform using std::filesystem
  instead of ifdefs.
- Add a bunch of missing float literal suffixes.
- Explicitly cast a bunch of types, especially in Lua bindings and GM
  tweaks.
- Gate more AISystem metrics vars to DEBUG builds.
- Remove a bunch of unused vars and lamba captures.
- Try removing redundant multiplayerReplicationFunctions instantiations.
- Fix broken LOG line in DEBUG build.

+++ src/gameGlobalInfo.cpp
@@ -63,7 +63,7 @@ GameGlobalInfo::GameGlobalInfo()
     registerMemberReplication(&allow_main_screen_long_range_radar);
     registerMemberReplication(&allow_main_screen_strategic_map);
     registerMemberReplication(&gm_control_code);
-    registerMemberReplication(&elapsed_time, 0.1);
+    registerMemberReplication(&elapsed_time, 0.1f);
     registerMemberReplication(&default_skybox);
 }
 
@@ -431,7 +431,7 @@ void GameGlobalInfo::destroy()
 }
 
 string GameGlobalInfo::getMissionTime() {
-    unsigned int seconds = gameGlobalInfo->elapsed_time;
+    unsigned int seconds = static_cast<unsigned int>(gameGlobalInfo->elapsed_time);
     unsigned int minutes = (seconds / 60) % 60;
     unsigned int hours = (seconds / 60 / 60) % 24;
     seconds = seconds % 60;
@@ -487,8 +487,8 @@ static int lettersToBlockNZ(const string& s, int& pos)
 string getSectorName(glm::vec2 position)
 {
     constexpr float sector_size = 20000;
-    int sector_x = floorf(position.x / sector_size) + 50;
-    int sector_y = floorf(position.y / sector_size) + 50;
+    int sector_x = static_cast<int>(floorf(position.x / sector_size)) + 50;
+    int sector_y = static_cast<int>(floorf(position.y / sector_size)) + 50;
 
     int block_x = sector_x >= 0 ? sector_x / 100 : (sector_x - 99) / 100;
     int local_col = ((sector_x % 100) + 100) % 100;
