+++ src/screenComponents/warpControls.cpp
@@ -13,31 +13,31 @@ GuiWarpControls::GuiWarpControls(GuiContainer* owner, string id)
 : GuiElement(owner, id)
 {
     // Build warp request slider.
-    slider = new GuiSlider(this, id + "_SLIDER", 1.0, 0.0, 0.0, [this](float value) {
+    slider = new GuiSlider(this, id + "_SLIDER", 1.0f, 0.0f, 0.0f, [this](float value) {
         // Round the slider value to an int.
-        int warp_level = value;
+        int warp_level = static_cast<int>(value);
 
         // Send a warp request command to our ship.
         if (my_spaceship)
             my_player_info->commandWarp(warp_level);
 
         // Set the slider value to the warp level.
-        slider->setValue(warp_level);
+        slider->setValue(static_cast<float>(warp_level));
     });
     slider->setPosition(0, 0, sp::Alignment::TopLeft)->setSize(50, GuiElement::GuiSizeMax);
-    slider->addSnapValue(0.0, 0.5);
-    slider->addSnapValue(1.0, 0.5);
+    slider->addSnapValue(0.0f, 0.5f);
+    slider->addSnapValue(1.0f, 0.5f);
 
     if (my_spaceship)
     {
         auto warp = my_spaceship.getComponent<WarpDrive>();
         // Set the slider's value to the current warp request.
         if (warp)
-            slider->setValue(warp->request);
+            slider->setValue(static_cast<float>(warp->request));
     }
 
     // Label the warp slider.
-    label = new GuiKeyValueDisplay(this, id + "_LABEL", 0.5, tr("slider", "Warp"), "0.0");
+    label = new GuiKeyValueDisplay(this, id + "_LABEL", 0.5f, tr("slider", "Warp"), "0.0");
     label->setTextSize(30)->setPosition(50, 0, sp::Alignment::TopLeft)->setSize(40, GuiElement::GuiSizeMax);
 
     // Prep the alert overlay.
@@ -52,13 +52,13 @@ void GuiWarpControls::onDraw(sp::RenderTarget& target)
     if (auto warp = my_spaceship.getComponent<WarpDrive>())
     {
         label->setValue(string(warp->current, 1));
-        slider->setValue(warp->request);
+        slider->setValue(static_cast<float>(warp->request));
 
-        if (slider->getRangeMin() != warp->max_level)
+        if (slider->getRangeMin() != static_cast<float>(warp->max_level))
         {
-            slider->setRange(warp->max_level, 0.0f);
+            slider->setRange(static_cast<float>(warp->max_level), 0.0f);
             for (int n = 0; n <= warp->max_level; n++)
-                slider->addSnapValue(n, 0.5f);
+                slider->addSnapValue(static_cast<float>(n), 0.5f);
         }
     }
 }
